Separated mailer and uploader
