+++ routes/apply/api.js
@@ -1,56 +1,8 @@
-//Api Routes
-var crypto = require('crypto');
-
-var mailer = require('nodemailer');
-var multer = require('multer');
-var mime = require('mime');
-var EmailTemplate = require('email-templates').EmailTemplate;
-
-var storage = multer.diskStorage({
-  dest: 'public/uploads/',
-  filename(req, file, cb) {
-    return crypto.pseudoRandomBytes(16, (err, raw) =>
-      cb(null, raw.toString('hex') + '.' + mime.extension(file.mimetype)));
-  }
-});
-
-var upload = multer({
-  storage,
-  //5MB file size
-  limits: {fileSize: 5 * 1024 * 1024}
-});
-
 module.exports = function(routes, config) {
-  // Node mailer
-  var transporter = mailer.createTransport({
-    host: config.MAIL_HOST,
-    port: config.MAIL_PORT,
-    secure: true,
-    auth: {
-      user: config.MAIL_USER,
-      pass: config.MAIL_PASS
-    }
-  });
+  var {referSender, confirmSender} = require('../../config/mailer')(config);
+  var upload = require('../../config/uploads')(config);
 
   var User = require('../../entities/users/model');
-
-  var confirmSender = transporter.templateSender(
-    new EmailTemplate('./views/emails/confirmation'),
-    {
-      from: {
-        name: 'SD Hacks Team',
-        address: process.env.MAIL_USER
-      }
-    });
-
-  var referSender = transporter.templateSender(
-    new EmailTemplate('./views/emails/refer'),
-    {
-      from: {
-        name: 'SD Hacks Team',
-        address: process.env.MAIL_USER
-      }
-    });
 
   var referTeammates = (user, req) =>
     // Queue up the referall emails
@@ -68,34 +20,6 @@ module.exports = function(routes, config) {
           }
         })
       )(referral));
-
-  routes.post('/upload', upload.single('resume'), function(req, res) {
-    var userError = function() {
-      res.status(400);
-      return res.json({'error': true});
-    };
-
-    if (!req.body.user_id || !req.file) {
-      return userError();
-    }
-
-    return User.findById(req.body.user_id, function(e, user) {
-      if (e || (user === null)) {
-        return userError();
-      }
-
-      return user.attach('resume', {path: req.file.path}, function(error) {
-        if (error) {
-          console.error(error);
-          console.error('Failed to upload new user resume');
-          userError();
-        }
-
-        user.save();
-        return res.json({'url': user.resume.url});
-      });
-    });
-  });
 
   routes.post('/register', upload.single('resume'), function(req, res) {
     var user = new User;
